requirements.txt
# baukit==0.0.1
# clean_fid==0.1.35
# clip==0.2.0
# clip_anytorch==2.5.2
diffusers==0.16.1
fairscale==0.4.13
# library==0.0.0
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
# Pillow==9.5.0
Pillow==10.0.0
safetensors==0.3.1
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.1
# util==0.0.1
